Addressed and incorporated review comments.
