+++ config/io/io_config.yaml
@@ -43,3 +43,5 @@
 "s3endpoints":
     "s3_io": "s3.seagate.com"
     "s3_auth": "iam.seagate.com:9443"
+
+"max_storage": 98.0  # Storage percentage should not exceed for IO to be run on cluster.
